Address some feedback, implement functionality in appropriate files, renames, etc.
